requirements.txt
adal==1.2.7
applicationinsights==0.11.10
argcomplete @ file:///tmp/build/80754af9/argcomplete_1618920853202/work
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
azure-common==1.1.27
azure-core==1.21.1
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==18.1.0
azure-mgmt-storage==19.0.0
azure-storage-blob==12.9.0
azure-storage-queue==12.1.6
azureml-automl-core==1.38.0
azureml-automl-runtime==1.38.0
azureml-contrib-automl-pipeline-steps==1.38.0
azureml-contrib-dataset==1.38.0
azureml-core==1.38.0
azureml-dataprep==2.26.0
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.2.0
azureml-dataset-runtime==1.38.0
azureml-defaults==1.38.0
azureml-explain-model==1.38.0
azureml-inference-server-http==0.4.2
azureml-interpret==1.38.0
azureml-mlflow==1.38.0
azureml-model-management-sdk==1.0.1b6.post1
azureml-pipeline==1.38.0
azureml-pipeline-core==1.38.0
azureml-pipeline-steps==1.38.0
azureml-sdk==1.38.0
azureml-telemetry==1.38.0
azureml-train-automl==1.38.0
azureml-train-automl-client==1.38.0
azureml-train-automl-runtime==1.38.0
azureml-train-core==1.38.0
azureml-train-restclients-hyperdrive==1.38.0
azureml-widgets==1.38.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
bleach==4.1.0
bokeh==2.4.2
boto==2.49.0
boto3==1.20.19
botocore==1.23.19
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.11.0
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262352237/work
configparser==3.7.4
contextlib2==21.6.0
cryptography @ file:///C:/ci/cryptography_1639414690059/work
cycler==0.11.0
Cython==0.29.14
dask==2.30.0
databricks-cli==0.16.2
dataclasses==0.6
debugpy @ file:///C:/ci/debugpy_1637091911212/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dill==0.3.4
distributed==2.30.1
distro==1.6.0
docker==5.0.3
dotnetcore2==2.1.22
entrypoints==0.3
fire==0.4.0
Flask==1.0.3
flatbuffers==2.0
fonttools==4.29.1
fsspec==2022.1.0
fusepy==3.0.1
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.26
gunicorn==19.9.0
HeapDict==1.0.1
humanfriendly==10.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.10.1
importlib-resources==5.4.0
inference-schema==1.3.0
interpret-community==0.24.1
interpret-core==0.2.7
ipykernel @ file:///C:/ci/ipykernel_1633527406033/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1635944342510/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.5
isodate==0.6.1
itsdangerous==2.0.1
jedi @ file:///C:/ci/jedi_1611315820849/work
jeepney==0.7.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
json-logging-py==0.2
JsonForm==0.0.2
jsonpickle==2.1.0
jsonschema==4.4.0
JsonSir==0.0.2
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-core @ file:///C:/ci/jupyter_core_1636537209059/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras2onnx==1.6.0
kiwisolver==1.3.2
knack==0.8.2
liac-arff==2.5.0
lightgbm==2.3.0
llvmlite==0.36.0
locket==0.2.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
menuinst==1.4.16
mistune==0.8.4
ml-wrappers==0.0.6
mlflow-skinny==1.23.0
msal==1.16.0
msal-extensions==0.3.1
msgpack==1.0.3
msrest==0.6.21
msrestazure==0.6.4
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
nimbusml==1.8.0
notebook==6.4.8
numba==0.53.1
numpy==1.18.5
oauthlib==3.1.1
onnx==1.7.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
onnxruntime==1.8.0
packaging==21.3
pandas==0.25.1
pandocfilters==1.5.0
paramiko==2.9.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partd==1.2.0
pathspec==0.9.0
patsy==0.5.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.0
pkginfo==1.8.2
pmdarima==1.1.1
portalocker==2.3.2
prometheus-client==0.13.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.3
psutil==5.8.0
py-cpuinfo==5.0.0
pyarrow==3.0.0
pyasn1==0.4.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==3.0.6
pyreadline==2.1
pyrsistent==0.18.1
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
Python-EasyConfig==0.1.7
pytz==2021.3
pywin32==227
pywinpty==2.0.1
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1638435182681/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
Resource==0.2.1
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016865685/work
ruamel.yaml==0.15.89
s3transfer==0.5.0
scikit-learn==0.22.1
scipy==1.1.0
SecretStorage==3.3.1
Send2Trash==1.8.0
shap==0.39.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
skl2onnx==1.4.9
sklearn-pandas==1.7.0
slicer==0.0.7
smart-open==1.9.0
smmap==5.0.0
sortedcontainers==2.4.0
statsmodels==0.10.2
tabulate==0.8.9
tblib==1.7.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
toolz==0.11.2
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.0.1
urllib3==1.26.7
waitress==2.0.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==0.16.1
widgetsnbextension==3.5.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
zict==2.0.0
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
